# server/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: VYBE API
  version: "1.0.0"
  summary: The multiplayer layer over any streaming service.
  description: |
    VYBE synchronises a server-authoritative **clock**, not a stream (ADR-002).
    You cannot sync playback of a service you do not control, so a room runs a
    virtual timeline and every timed feature fires against it:

        t_room = (server_now − anchor) + offset

    Three conventions apply everywhere and are not repeated per-endpoint:

    * **Errors** are `application/problem+json` with a stable `code` and a
      `traceId`. Branch on `code`.
    * **Pagination** is opaque cursors, never offsets. Sending `?offset=`,
      `?page=`, or `?skip=` is refused with `OFFSET_PAGINATION_UNSUPPORTED`
      rather than ignored.
    * **Mutations** require an `Idempotency-Key`. A retry with the same key
      replays the first response and carries `Idempotent-Replay: true`.
  license:
    name: Proprietary
  contact:
    name: eyadsyam
servers:
  - url: http://localhost:8080
    description: local development
  - url: https://api.vybe.app
    description: production
...
security:
  - bearerAuth: []
paths:
...
  /v1/rooms:
    get:
      tags: [rooms]
      summary: Rooms the caller hosts or has joined
      operationId: listRooms
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: |
            Newest first. `items` is always an array, never null.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Room" }
                  nextCursor:
                    type: string
                    description: Absent on the last page. Echo it back unmodified.
        "400":
          description: |
            A malformed cursor (`CURSOR_INVALID`), a bad limit
            (`LIMIT_INVALID`), or an attempt at offset pagination
            (`OFFSET_PAGINATION_UNSUPPORTED`).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401": { $ref: "#/components/responses/Unauthorized" }

    post:
      tags: [rooms]
      summary: Open a room
      description: |
        Creates a room in `LOBBY` and returns its join code. Capacity comes
        from the HOST's entitlement tier (4 free, 8 plus) and is a property of
        the room from then on — a free-tier guest can join a plus host's room.
      operationId: createRoom
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRoomRequest" }
      responses:
        "201":
          description: The room, including its join code and share URL.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/IdempotencyInFlight" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422":
          description: |
            Validation failed, or `CONTENT_NOT_FOUND` — no such title in the
            catalogue.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "503":
          description: A join code could not be allocated. Retry.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/rooms/join:
    post:
      tags: [rooms]
      summary: Join a room by code
      description: |
        Join codes are Crockford base32 and are decoded leniently, because a
        code is read aloud over a call and typed by somebody else: case is
        ignored, hyphens and spaces are stripped, and `I`/`l` decode to `1`
        while `O`/`o` decode to `0`. `U` is **excluded** from the alphabet and
        is rejected rather than mapped — unlike the others there is no digit it
        is plausibly a misreading of, and silently mapping it would resolve to
        a different room.

        A malformed code and an unknown code both return `ROOM_NOT_FOUND`.
        Distinguishing them would tell somebody enumerating codes that their
        alphabet is right.
      operationId: joinRoom
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [joinCode]
              properties:
                joinCode:
                  type: string
                  examples: ["K7X2QP", "k7x-2qp"]
      responses:
        "200":
          description: Joined. The response includes the current participants.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No such live room — or the code was malformed.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: |
            `ROOM_FULL`, `ALREADY_JOINED`, or `ROOM_ENDED`.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/rooms/{roomId}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      tags: [rooms]
      summary: Read a room
      description: |
        Membership is required. Possession of a join code is **not** access
        (FR-14) — a code shared into a group chat months ago must not remain a
        permanent key.

        A non-member receives `404 ROOM_NOT_FOUND`, not 403. Telling a stranger
        "that room exists but you are not in it" confirms the id, which is
        exactly what somebody probing ids wants to learn.
      operationId: getRoom
      responses:
        "200":
          description: The room and its participants.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/RoomNotFound" }
  /v1/rooms/{roomId}/leave:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [rooms]
      summary: Leave a room
      description: |
        If the HOST leaves, the longest-tenured connected participant is
        promoted and a `HOST_CHANGED` event is emitted. If nobody is left, the
        room ends immediately rather than waiting for the reaper — an empty
        lobby that still resolves by join code is a room a straggler can walk
        into alone.

        The response omits `joinCode`: the caller is no longer a member.
      operationId: leaveRoom
      responses:
        "200":
          description: Left. The room as it now stands, without the join code.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Not a participant.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404": { $ref: "#/components/responses/RoomNotFound" }
  /v1/rooms/{roomId}/end:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [rooms]
      summary: End a room
      description: |
        Host only. **Idempotent**: ending an already-ended room succeeds and
        emits no second `ROOM_ENDED`, so a host double-tapping on a slow
        connection does not see a failure for something that worked.
      operationId: endRoom
      responses:
        "200":
          description: Ended.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotTheHost" }
        "404": { $ref: "#/components/responses/RoomNotFound" }
  /v1/rooms/{roomId}/transition:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [rooms]
      summary: Drive the room state machine
      description: |
        Host only. The legal transitions (FR-15):

            LOBBY   --ARM-->       READY
            READY   --START-->     PLAYING
            READY   --CANCEL-->    LOBBY
            PLAYING --REANCHOR-->  READY
            any     --END-->       ENDED

        An illegal transition returns `409 ILLEGAL_TRANSITION` with `fromState`
        and `attemptedEvent` as extension members, so a client can tell the
        user what the room's state actually is rather than guessing.
      operationId: transitionRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event]
              properties:
                event:
                  type: string
                  enum: [ARM, START, REANCHOR, CANCEL, END]
      responses:
        "200":
          description: Transitioned.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotTheHost" }
        "404": { $ref: "#/components/responses/RoomNotFound" }
        "409":
          description: Not available from the room's current state.
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Problem"
                  - type: object
                    properties:
                      fromState: { $ref: "#/components/schemas/RoomState" }
                      attemptedEvent: { type: string }
              example:
                type: https://vybe.app/problems/illegal-transition
                title: Illegal Transition
                status: 409
                code: ILLEGAL_TRANSITION
                detail: Cannot START from LOBBY.
                traceId: 01H5exitzSg1J3dCy4NnS8jG
                fromState: LOBBY
                attemptedEvent: START
        "422": { $ref: "#/components/responses/ValidationFailed" }

  # ---------------------------------------------------------------------------
  # Realtime
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An Ed25519-signed JWT, 15-minute lifetime. The scheme name is
        case-insensitive per RFC 7235.

        The token must be in the `Authorization` header. Passing it as
        `?access_token=`, `?token=`, or `?jwt=` is refused with
        `TOKEN_IN_QUERY` even when a valid header is also present.
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        8–255 printable ASCII characters, unique per logical operation
        (FR-57). Keys are scoped to the caller, so two users may use the same
        key without colliding.

        A retry with the same key AND the same body replays the original
        response with `Idempotent-Replay: true`. The same key with a
        DIFFERENT body is a client bug and returns
        `422 IDEMPOTENCY_KEY_REUSED` rather than silently replaying the wrong
        response.
      schema:
        type: string
        minLength: 8
        maxLength: 255
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        Opaque. Echo back the previous page's `nextCursor` unmodified; do not
        construct one.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: |
        Not authenticated. `TOKEN_EXPIRED` specifically means "refresh and
        retry silently"; every other code here means "send the user to login".
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationFailed:
      description: |
        One or more fields are invalid. `errors[]` names them. Those `detail`
        strings are diagnostics for a developer reading logs, NOT display
        copy — FR-61 requires user-facing text to come from the client's own
        .arb files, so render from `code`.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    UnsupportedMediaType:
      description: The body must be `application/json`.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    IdempotencyInFlight:
      description: |
        A request with this key is still being processed. Retry shortly —
        this is not a failure.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    RoomNotFound:
      description: |
        No such room, **or** the caller is not a member of it. The two are
        deliberately indistinguishable.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotTheHost:
      description: Only the host can do that.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: RFC 9457. The only error shape in the API.
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri
          description: A documentation anchor, not a dereferenceable endpoint.
        title: { type: string, description: Human-readable. Free to be reworded; do not match on it. }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          description: |
            **Branch on this.** Stable, SCREAMING_SNAKE_CASE, and part of the
            contract — renaming one is a breaking change. `status` is too
            coarse: five different situations share 409.
        traceId:
          type: string
          description: |
            Echoes the client's `X-Trace-Id` when supplied, so a user
            reporting "it failed" can be joined to server logs by the id their
            app already recorded (§14.2).
        errors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
    FieldError:
      type: object
      required: [field, code, detail]
      properties:
        field: { type: string }
        code: { type: string }
        detail:
          type: string
          description: A developer diagnostic, not display copy. See ValidationFailed.
...
    RoomState:
      type: string
      enum: [LOBBY, READY, PLAYING, ENDED]
    CreateRoomRequest:
      type: object
      required: [contentId]
      properties:
        contentId: { type: string, format: uuid }
        title: { type: string, maxLength: 120, description: Counted in characters, not bytes. }
        visibility: { type: string, enum: [private, public], default: private }
        syncMode:
          type: string
          enum: [COMPANION, MANUAL]
          default: COMPANION
          description: COMPANION is the mode ADR-002 exists for.
    Participant:
      type: object
      required: [userId, isHost, connected, joinedAt]
      properties:
        userId: { type: string, format: uuid }
        isHost: { type: boolean }
        connected:
          type: boolean
          description: Live socket presence, from the hub rather than the database.
        joinedAt: { type: string, format: date-time }
    Room:
      type: object
      required: [id, contentId, hostUserId, visibility, state, syncMode, maxParticipants, currentSeq, createdAt, serverTime]
      properties:
        id: { type: string, format: uuid }
        contentId: { type: string, format: uuid }
        hostUserId: { type: string, format: uuid }
        joinCode:
          type: string
          pattern: "^[0-9A-HJKMNP-TV-Z]{6}$"
          description: |
            Members only. Absent from any response a non-member can obtain,
            and from the `leave` response — it is the credential that admits
            somebody to the room.
        shareUrl:
          type: string
          format: uri
          description: |
            An https Universal Link, never a custom scheme: any app can claim
            a custom scheme, whereas a Universal Link is bound to a domain an
            attacker does not control (§1.9 L4).
        visibility: { type: string, enum: [private, public] }
        state: { $ref: "#/components/schemas/RoomState" }
        syncMode: { type: string, enum: [COMPANION, MANUAL] }
        title: { type: string }
        maxParticipants: { type: integer, description: From the HOST's tier. }
        currentSeq:
          type: integer
          format: int64
          description: |
            The room's position in its event log. A reconnecting client
            compares its own `lastSeq` against this to decide whether it needs
            a resync at all.
        anchorServerTime:
          type: string
          format: date-time
          description: ADR-002's anchor. Present once the room has started.
        anchorOffsetMs: { type: integer, format: int64 }
        reanchorCount: { type: integer }
        createdAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time }
        endReason:
          type: string
          enum: [host_ended, reaper_abandoned, moderation, error]
        participants:
          type: array
          items: { $ref: "#/components/schemas/Participant" }
        serverTime:
          type: string
          format: date-time
          description: |
            The server's clock at the moment of the response. Present on every
            room payload so a client always has a fresh reference for ADR-002
            without a second round trip.
...